+++ src/main/java/gtclassic/GTMod.java
@@ -30,6 +30,7 @@ import gtclassic.common.proxy.GTProxyCommon;
 import gtclassic.common.recipe.GTRecipe;
 import gtclassic.common.recipe.GTRecipeIterators;
 import gtclassic.common.recipe.GTRecipeMods;
+import gtclassic.common.tile.GTTileBath;
 import gtclassic.common.tile.GTTileCentrifuge;
 import gtclassic.common.tile.GTTileDisassembler;
 import gtclassic.common.tile.GTTileMagicEnergyConverter;
@@ -58,7 +59,7 @@ public class GTMod {
 
 	public static final String MODID = "gtclassic";
 	public static final String MODNAME = "GregTech Classic";
-	public static final String MODVERSION = "1.1.5";
+	public static final String MODVERSION = "1.1.6";
 	public static final String DEPENDS = "required-after:ic2;required-after:ic2-classic-spmod;before:gtc_expansion@[0.0.9,);after:twilightforest@[3.9.984,)";
 	public static final CreativeTabs creativeTabGT = new GTCreativeTab(MODID);
 	@SidedProxy(clientSide = MODID + ".common.proxy.GTProxyClient", serverSide = MODID + ".common.proxy.GTProxyServer")
@@ -96,6 +97,7 @@ public class GTMod {
 		GTMaterialElement.init();
 		GTRecipeIterators.init();
 		GTTileCentrifuge.init();
+		GTTileBath.init();
 		GTTileUUMAssembler.init();
 		GTTileMagicEnergyConverter.init();
 		GTRecipe.initShapeless();